fixes test_affined

+++ tests/test_affined.py
@@ -31,7 +31,7 @@ for p in TEST_NDARRAYS:
         TESTS.append(
             [
                 dict(keys="img", padding_mode="zeros", spatial_size=(-1, 0), device=device, dtype=None),
-                {"img": p(np.arange(9).reshape((1, 3, 3)))},
+                {"img": p(np.arange(9, dtype=float).reshape((1, 3, 3)))},
                 p(np.arange(9).reshape(1, 3, 3)),
             ]
         )
